misc: remove playwright container
The intention was to utilize playwright to enable service-monitor to render graphical representations of changes in the watched website.

This proved to be less useful than desired, and the browserless/chrome image eats up a fair bit of RAM.

Removing container so resources aren't tied up with the unused service.